requirements.txt
python-dotenv>=1.2.0
-e git+https://github.com/Azure/PyRIT@7ff47ebd165441e54607cea5d6fdc7999427b816#egg=pyrit
openai>=2.11.0
anthropic>=0.34.0
google-generativeai>=0.8.0
transformers>=4.57.0
huggingface-hub>=0.36.0
torch>=2.5.1
torchvision>=0.20.1
accelerate>=1.0.0